requirements.txt
fastapi
uvicorn
psycopg2-binary
openai
python-dotenv
streamlit
requests
pydantic
pytest
httpx
anthropic
google
google-genai
